Version 1.3 camera overlay, new prefs, and fast deletion

+++ src/wei/mark/delete/util/Utils.java
@@ -4,7 +4,6 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-import wei.mark.delete.ShareToDeleteActivity;
 import android.app.AlertDialog;
 import android.content.ContentResolver;
 import android.content.Context;
@@ -18,7 +17,6 @@ import android.provider.BaseColumns;
 import android.provider.MediaStore;
 import android.util.Log;
 import android.widget.ImageView;
-import android.widget.Toast;
 
 public class Utils {
 
@@ -26,6 +24,9 @@ public class Utils {
 
 	public static Bitmap createThumbnail(Bitmap imageBitmap,
 			int thumbnailWidth, int thumbnailHeight) {
+		if (imageBitmap == null)
+			return null;
+
 		Float width = new Float(imageBitmap.getWidth());
 		Float height = new Float(imageBitmap.getHeight());
 		Float ratio = width / height;
@@ -46,15 +47,11 @@ public class Utils {
 		}
 	}
 
-	public static void delete(final ShareToDeleteActivity context,
-			final String path, final Uri uri, boolean prompt,
+	public static void delete(final Context context, final Uri uri,
 			final DeleteCallback callback) {
-
-		prompt = prompt
-				|| PreferenceManager.getDefaultSharedPreferences(context)
-						.getBoolean("prompt", false);
+		boolean prompt = PreferenceManager.getDefaultSharedPreferences(context)
+				.getBoolean("prompt", false);
 		if (prompt) {
-
 			ImageView image = new ImageView(context);
 			Bitmap bitmap;
 			try {
@@ -69,13 +66,12 @@ public class Utils {
 
 			AlertDialog.Builder alert = new AlertDialog.Builder(context);
 			alert.setTitle("What do you want to do?")
-					.setMessage(path)
+					.setMessage(getPath(context, uri))
 					.setPositiveButton("Delete", new OnClickListener() {
 
 						@Override
 						public void onClick(DialogInterface dialog, int which) {
-							callback.done(deleteWithoutPrompt(context, path,
-									uri, getId(context, uri)));
+							callback.done(deleteWithoutPrompt(context, uri));
 							return;
 						}
 					}).setNegativeButton("Keep", new OnClickListener() {
@@ -90,50 +86,87 @@ public class Utils {
 					.setIcon(android.R.drawable.ic_menu_delete).show();
 		} else {
 			// no prompt
-			callback.done(deleteWithoutPrompt(context, path, uri,
-					getId(context, uri)));
+			callback.done(deleteWithoutPrompt(context, uri));
 		}
 	}
 
-	private static boolean deleteWithoutPrompt(ShareToDeleteActivity context,
-			String path, Uri uri, int mediaId) {
-		File file = new File(path);
-		if (file.canWrite()) {
-			if (file.delete()) {
-				ContentResolver cr = context.getContentResolver();
-				int deleted = 0;
-				if ((deleted = cr.delete(
-						MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
-						BaseColumns._ID + "=" + mediaId, null)) == 1) {
-					Toast.makeText(context, "Deleted: " + path,
-							Toast.LENGTH_SHORT).show();
-					return true;
+	private static boolean deleteWithoutPrompt(Context context, Uri uri) {
+		String path = getPath(context, uri);
+		if (path != null) {
+			File file = new File(path);
+			if (file.canWrite()) {
+				if (file.delete()) {
+					ContentResolver cr = context.getContentResolver();
+					int deleted = 0;
+					int id;
+					try {
+						id = getId(context, uri);
+					} catch (ShareToDeleteException e) {
+						Utils.fail("Deleted "
+								+ deleted
+								+ " from storage but not from Content Provider.");
+						return false;
+					}
+					if ((deleted = cr.delete(
+							MediaStore.Images.Media.EXTERNAL_CONTENT_URI,
+							BaseColumns._ID + "=" + id, null)) == 1) {
+						return true;
+					} else {
+						Utils.fail("Deleted " + deleted
+								+ " rows from Content Provider.");
+					}
 				} else {
-					Utils.fail("Deleted " + deleted
-							+ " rows from Content Provider.");
+					Utils.fail("Failed to delete: " + path);
 				}
 			} else {
-				Utils.fail("Failed to delete: " + path);
+				Utils.fail("Cannot write to path: " + path);
 			}
 		} else {
-			Utils.fail("Cannot write to path: " + path);
+			Utils.fail("File not found: " + path);
 		}
 
 		return false;
 	}
 
-	public static int getId(Context context, Uri uri) {
+	public static int getId(Context context, Uri uri)
+			throws ShareToDeleteException {
+		Log.d("Share To Delete", "Delete requested. Uri: " + uri);
 		String[] projection = { MediaStore.Images.Media._ID };
 		Cursor cursor = context.getContentResolver().query(uri, projection,
 				null, null, null);
 		int column_index = cursor
 				.getColumnIndexOrThrow(MediaStore.Images.Media._ID);
 
-		cursor.moveToFirst();
-		return cursor.getInt(column_index);
+		if (cursor.moveToFirst()) {
+			return cursor.getInt(column_index);
+		}
+
+		throw new ShareToDeleteException("Image does not exist.");
+	}
+
+	// http://stackoverflow.com/questions/5548645/get-thumbnail-uri-path-of-the-image-stored-in-sd-card-android
+	public static String getPath(Context context, Uri uri) {
+		String[] projection = { MediaStore.Images.Media.DATA };
+		Cursor cursor = context.getContentResolver().query(uri, projection,
+				null, null, null);
+		int column_index = cursor
+				.getColumnIndexOrThrow(MediaStore.Images.Media.DATA);
+
+		if (cursor.moveToFirst()) {
+			return cursor.getString(column_index);
+		}
+		return null;
 	}
 
 	public interface DeleteCallback {
 		void done(boolean success);
+	}
+
+	public static class ShareToDeleteException extends Exception {
+		private static final long serialVersionUID = -3175598643034955186L;
+
+		public ShareToDeleteException(String message) {
+			super(message);
+		}
 	}
 }
